Renamed the page_title module to page_header. Added some clippy lints fixes
